requirements.txt
autopep8==1.6.0
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cymem==2.0.6
Cython==0.29.23
gensim==4.1.2
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.0
murmurhash==1.0.6
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathy==0.6.1
preshed==3.0.6
pycodestyle==2.8.0
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.5
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.63.0
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
